+++ SourceryFramework/Models/Struct.swift
@@ -15,6 +15,7 @@ public final class Struct: Type {
     /// Returns "struct"
     public override var kind: String { return "struct" }
 
+    /// :nodoc:
     public override init(name: String = "",
                   parent: Type? = nil,
                   accessLevel: AccessLevel = .internal,