+++ SourceryFramework/Models/Typealias.swift
@@ -1,6 +1,7 @@
 import Foundation
 
 // sourcery: skipJSExport
+/// :nodoc:
 public final class Typealias: NSObject, Typed, SourceryModel {
     // New typealias name
     public let aliasName: String