+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Paths in config file are now relative to config file path by default, absolute paths should start with `/`
 - Improved logging and error reporting, added `--quiet` CLI option, added runtime errors for using invalid types in `implementing` and `inheriting`
+
+### Internal changes
+
+- Improved `AutoCases.stencil` to use `let` instead of computed `var`
 - Add the `lowerFirst` filter for Stencil templates.
 
 ## 0.6.0